Refactor renderer state updates
