+++ src/heart/renderers/l_system/provider.py
@@ -9,6 +9,29 @@ from heart.peripheral.core.providers import ObservableProvider
 from heart.renderers.l_system.state import LSystemState
 
 
+def _update_grammar(grammar: str) -> str:
+    expanded = ""
+    for char in grammar:
+        if char == "X":
+            expanded += "F+[[X]-X]-F[-FX]+X"
+        elif char == "F":
+            expanded += "FF"
+    return expanded
+
+
+def _advance_state(
+    state: LSystemState, *, dt_ms: float, update_interval_ms: float
+) -> LSystemState:
+    accumulated = state.time_since_last_update_ms + dt_ms
+    grammar = state.grammar
+
+    while accumulated >= update_interval_ms:
+        grammar = _update_grammar(grammar)
+        accumulated -= update_interval_ms
+
+    return LSystemState(grammar=grammar, time_since_last_update_ms=accumulated)
+
+
 class LSystemStateProvider(ObservableProvider[LSystemState]):
     def __init__(
         self,
@@ -24,10 +47,11 @@ class LSystemStateProvider(ObservableProvider[LSystemState]):
             ops.share(),
         )
 
-        initial_state = LSystemState.initial()
+        initial_state = LSystemState()
 
         def advance(state: LSystemState, clock: Clock) -> LSystemState:
-            return state.advance(
+            return _advance_state(
+                state,
                 dt_ms=float(clock.get_time()),
                 update_interval_ms=self._update_interval_ms,
             )
