+++ src/heart/renderers/max_bpm_screen/state.py
@@ -1,5 +1,3 @@
-from __future__ import annotations
-
 from dataclasses import dataclass
 
 
@@ -8,7 +6,3 @@ class AvatarBpmRendererState:
     sensor_id: str | None
     bpm: int | None
     avatar_name: str | None
-
-    @classmethod
-    def initial(cls) -> "AvatarBpmRendererState":
-        return cls(sensor_id=None, bpm=None, avatar_name=None)
