+++ src/heart/renderers/porthole_window/provider.py
@@ -23,7 +23,9 @@ class PortholeWindowStateProvider(ObservableProvider[PortholeWindowState]):
 
         def advance_state(state: PortholeWindowState, clock: Clock) -> PortholeWindowState:
             delta_seconds = max(clock.get_time(), 0) / 1000.0
-            return state.advance(delta_seconds)
+            return PortholeWindowState(
+                elapsed_seconds=state.elapsed_seconds + delta_seconds
+            )
 
         return (
             self._peripheral_manager.game_tick.pipe(
