+++ src/heart/renderers/text/state.py
@@ -1,9 +1,4 @@
-from __future__ import annotations
-
 from dataclasses import dataclass
-from functools import cached_property
-
-import pygame
 
 from heart.display.color import Color
 from heart.peripheral.switch import SwitchState
@@ -18,10 +13,3 @@ class TextRenderingState:
     color: Color
     x_location: int | None
     y_location: int | None
-
-    @cached_property
-    def font(self) -> pygame.font.Font:
-        return pygame.font.SysFont(
-            self.font_name,
-            self.font_size,
-        )
